user api created with authentication for the tasks

+++ server/controllers/taskController.js
@@ -1,12 +1,12 @@
 const asyncHandler = require('express-async-handler');
-
 const Task = require('../models/taskModel')
+const User = require('../models/userModel')
 
 // @desc Get task
 // @route GET /api/task
 // @access Private
 const getTask =  asyncHandler(async (req, res) => {
-    const tasks = await Task.find()
+    const tasks = await Task.find({ user: req.user })
     res.status(200).json(tasks)
 })
 
@@ -20,7 +20,8 @@ const setTask = asyncHandler(async (req, res) => {
     }
 
     const task = await Task.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(task)
@@ -33,8 +34,21 @@ const updateTask = asyncHandler(async (req, res) => {
     const task = await Task.findById(req.params.id)
 
     if(!task){
-        res.status(400)
+        res.status(401)
         throw new Error('Task not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //checking the logged in user matches the task user
+    if(task.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, {
@@ -53,6 +67,19 @@ const deleteTask = asyncHandler(async (req, res) => {
     if(!task){
         res.status(400)
         throw new Error('Task not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //checking the logged in user matches the task user
+    if(task.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await task.deleteOne()
